refactor: create createEventButton in entities
