+++ src/entities/createEventButton/style.css
